fix: canExportMetadata
